Refactor medicine module structure and routing
Moved medicine-related functions, types, and schemas into dedicated 'medicine' subdirectories for better organization. Introduced MedicineController to handle logic, updated routes to use controller methods, and simplified main.ts routing. This improves code maintainability and clarity.
