+++ functions/medicine/araia/config.ts
