+++ types/functions/medicine/searchProduct/index.ts
