+++ types/functions/medicine/verifyStore/index.ts
